Add texture to material creator (editor tool)
The tool allows to select a input folder, which has textures inside and
an output folder that is used to store materials created from the textures
The material names will be the same as the texture names, unless
a custom prefix and/or suffix for the material names is provided.
A template material can be selected, so that all created materials
will copy the template material attributes. Optionally, it is possible
to override the shader. Right now, the texture is used as entry for
the albedo map in the material (assigned to main texture property)
Other maps are not automatically added.